initial of delete article
