+++ Ptt/API/Models/GeneralResponse.swift
@@ -1,0 +1,15 @@
+//
+//  GeneralResponse.swift
+//  Ptt
+//
+//  Created by 陳建佑 on 2023/12/9.
+//  Copyright © 2023 Ptt. All rights reserved.
+//
+
+import Foundation
+
+extension APIModel {
+    struct GeneralResponse: Codable {
+        let success: Bool
+    }
+}
